requirements.txt
streamlit
langchain
python-dotenv
langchain-openai
langchain-community
